// src/app/services/models/kyc-verification.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface KycVerification {
  addressDocumentUrl?: string;
  addressRejectionReason?: string;
  addressVerificationStatus?: 'PENDING' | 'APPROVED' | 'REJECTED' | 'UNVERIFIED';
  expiryDate?: string;
  id?: number;
  idCountry?: string;
  idDocumentUrl?: string;
  idNumber?: string;
  idRejectionReason?: string;
  idType?: string;
  idVerificationStatus?: 'PENDING' | 'APPROVED' | 'REJECTED' | 'UNVERIFIED';
  userId?: string;
}